safari-600.7-branch/LayoutTests/platform/ios-sim/fast/dom/vertical-scrollbar-when-dir-change-expected.txt

original scroll y for RTL page : Success
vertical scroll y for RTL page : Success
original scroll y for LTR page : Failure, actual: 3310; expected: 0
vertical scroll y for LTR page : Failure, actual: 0; expected: 100
